Trasfer pictures to a local directory
The pictures in the recipes were linked to an external URL, in order to prevent the unlikely scenario of images breaking if the external page were to take down their image, I have downloaded the image files and stored them in a local directory. This required that I also edit the src in the <img> elements inside the recipe pages.
* Create ./images
* Store image files in this new directory
* Edit drunken_noodles.html so that <img> points to ../images/drunken_noodles.jpeg
* Edit lasagna.html so that <img> points to ../images/lasagna.jpeg
* Edit strawberry_muffins.html so that <img> points to ../images/strawberry_muffins.jpeg
* Also, I added an alt attribute to the above <img> elements